requirements.txt
python-dotenv==1.2.1
requests==2.32.5
feedparser==6.0.12
yfinance==1.2.0
pykrx==1.2.4
python-telegram-bot==22.6
schedule==1.2.2
google-generativeai==0.8.6
